New instances fixed memory issue in Combo

+++ include/combo/combo.h
@@ -24,7 +24,7 @@ extern "C" {
    ====================================================================== */
 
 typedef int boolean; /* logical variable         */
-typedef long ntype;   /* number of states/items   */
+typedef int ntype;   /* number of states/items   */
 typedef long itype;   /* item profits and weights */
 typedef long stype;   /* sum of profit or weight  */
